+++ src/gwt/src/org/rstudio/core/client/ElementIds.java
@@ -201,6 +201,11 @@ public class ElementIds
    public final static String NEW_PROJECT_SOURCE_FILES = "source_files";
    public final static String NEW_PROJECT_NEW_SESSION = "new_session";
 
+   public final static String SIDEBAR_COLUMN_SPLITTER = "sidebar_column_splitter";
+   public final static String SOURCE_COLUMN_SPLITTER = "source_column_splitter";
+   public final static String MIDDLE_COLUMN_SPLITTER = "middle_column_splitter";
+   public final static String CUSTOMIZE_PANES_BUTTON = "customize_panes";
+
    // AskSecretDialog
    public final static String ASK_SECRET_TEXT = "ask_secret_text";
    public static String getAskSecretText() { return getElementId(ASK_SECRET_TEXT); }