+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/model/UserPrefsAccessor.java
@@ -3875,6 +3875,18 @@ public class UserPrefsAccessor extends Prefs
    public final static String CONSOLE_HIGHLIGHT_CONDITIONS_ERRORS = "errors";
    public final static String CONSOLE_HIGHLIGHT_CONDITIONS_NONE = "none";
 
+   /**
+    * Whether to show the experimental Chat UI
+    */
+   public PrefValue<Boolean> showChatUi()
+   {
+      return bool(
+         "show_chat_ui",
+         _constants.showChatUiTitle(), 
+         _constants.showChatUiDescription(), 
+         false);
+   }
+
    public void syncPrefs(String layer, JsObject source)
    {
       if (source.hasKey("run_rprofile_on_resume"))
@@ -4415,6 +4427,8 @@ public class UserPrefsAccessor extends Prefs
          projectUserDataDirectory().setValue(layer, source.getString("project_user_data_directory"));
       if (source.hasKey("console_highlight_conditions"))
          consoleHighlightConditions().setValue(layer, source.getString("console_highlight_conditions"));
+      if (source.hasKey("show_chat_ui"))
+         showChatUi().setValue(layer, source.getBool("show_chat_ui"));
    }
    public List<PrefValue<?>> allPrefs()
    {
@@ -4688,6 +4702,7 @@ public class UserPrefsAccessor extends Prefs
       prefs.add(reformatOnSave());
       prefs.add(projectUserDataDirectory());
       prefs.add(consoleHighlightConditions());
+      prefs.add(showChatUi());
       return prefs;
    }