+++ src/gwt/src/org/rstudio/studio/client/workbench/ui/UIConstants.java
@@ -26,4 +26,7 @@ public interface UIConstants extends com.google.gwt.i18n.client.Messages {
     String cannotAddColumnText();
     String cannotAddMoreColumnsText(int maxColumnCount);
     String closeText();
+    String configurePanesButtonText();
+    String noTabsAssignedText();
+    String sidebarTitleText();
 }